Initial Implemntation of Sprint 2
git-tfs-id: [http://tfs2012.dowjones.net:8080/tfs/DefaultCollection]$/Dow Jones Infrastructure;C103555

+++ ServicesApplications/trunk/src/WidgetManagement/WidgetManagerApplication/du/EditWidget.aspx.cs
@@ -73,7 +73,7 @@ namespace EMG.widgets.ui.du
                 {
                     WorkspaceManager workspaceManager = new WorkspaceManager(SessionData.Instance().SessionBasedControlDataEx, SessionData.Instance().InterfaceLanguage);
                     // Find if there is a corresponding asset else perform a create
-                    int temp;
+                    long temp;
                     if (workspaceManager.IsWorkspaceIdAssociatedWithDissemenatedWidget(m_WidgetManagementDTO.assetIds[0], out temp))
                     {
                         m_WidgetManagementDTO.widgetId = temp.ToString();
